urlshortener_app/requirements.txt
# Core
pytz==2020.1
Django==3.0.8
django-crispy-forms==1.9.2
django-configurations==2.2
gunicorn==20.0.4

psycopg2-binary==2.8.5

# Developer Tools
ipdb==0.13.3
ipython==7.14.0
flake8==3.8.2

# Testing
mock==4.0.2
factory-boy==2.12.0
coverage==5.1
